[trainer, cfg, tests] refactor: consolidate distillation package into five cohesive modules
The distillation package was split into 14 files that exceeded the
repository's convention of grouping one responsibility per file with a
registry (cf. diffusion_algos.py holding all losses + advantage estimators +
registry). Merge the declarative strata:

- contracts.py absorbs role_runtime/export/score_providers/checkpoint (all
  data-contract and validation strata).
- recipes.py absorbs registry/objectives/rollout_strategies (all the
  declarative dispatch: registered objectives, rollout strategies, recipes).
- control_plane.py absorbs phase_executor (executor protocol + CPU fakes).

Net -177 lines. Public symbols are unchanged in name and behavior; import
paths for every symbol now resolve through the retained modules.

Clarify in math.py why it is its own module: it holds the only executable
equations (contracts = data types, recipes = declarations, math = equations),
so it can be unit-tested as algebraic identities without a plan/executor.

No model pipeline, Ray worker, FSDP, or GPU runtime changed. All 97 PR 1 CPU
tests and the full trainer/diffusion suite (334) pass.

AI assistance was used for this change.
